[refactor] 함수 정리, 패키지 정리
