+++ app/src/main/java/com/squirtles/musicroad/common/picklist/components/EditModeAction.kt
@@ -1,4 +1,4 @@
-package com.squirtles.musicroad.common.picklist
+package com.squirtles.musicroad.common.picklist.components
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Edit
