+++ app/src/main/java/com/squirtles/musicroad/common/picklist/components/EditModeBottomButton.kt
@@ -1,4 +1,4 @@
-package com.squirtles.musicroad.common.picklist
+package com.squirtles.musicroad.common.picklist.components
 
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxHeight
